signal_peptide_feature_table.py
Generates a csv table that for each accesstion number has a label whether
that signal sequence is sensitive or not. Then a sliding window is applied
of length [9:] to calculate min and max values for each of the scales that
are stored in files under scale_tables folder. Then for each scale min and
max values are stored in the table.